add new kind of key
