+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         tools:targetApi="tiramisu">
 
         <service
-            android:name=".IMEService"
+            android:name=".keyboard.IMEService"
             android:configChanges="uiMode"
             android:label="Example Comopose IME"
             android:permission="android.permission.BIND_INPUT_METHOD"
