+++ app/src/main/java/com/example/customkeyboard/keyboard/ComposeKeyboardView.kt
@@ -1,4 +1,4 @@
-package com.example.customkeyboard
+package com.example.customkeyboard.keyboard
 
 import android.content.Context
 import android.os.Build
