+++ app/src/main/java/com/example/customkeyboard/screens/ColorScreen.kt
@@ -31,9 +31,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
-import com.example.customkeyboard.IMEService
 import com.example.customkeyboard.R
 import com.example.customkeyboard.inputMethodManager
+import com.example.customkeyboard.keyboard.IMEService
 import com.github.skydoves.colorpicker.compose.ColorEnvelope
 import com.github.skydoves.colorpicker.compose.HsvColorPicker
 import com.github.skydoves.colorpicker.compose.rememberColorPickerController
